// src/types/locationSuggestions.type.ts
export interface ILocationSuggestions {
  name: string
  local_names: LocalNames
  lat: number
  lon: number
  country: string
  state?: string
}

export interface LocalNames {
  es?: string
  oc?: string
  pl?: string
  uk?: string
  it?: string
  ru: string
  de?: string
  fr?: string
  be?: string
  en: string
  ar?: string
  cs?: string
  sh?: string
  gl?: string
  eu?: string
  lb?: string
  he?: string
  ga?: string
  et?: string
  sw?: string
  ms?: string
  ha?: string
  ia?: string
  am?: string
  fy?: string
  mg?: string
  sn?: string
  sq?: string
  lt?: string
  cy?: string
  an?: string
  bg?: string
  lv?: string
  ko?: string
  fi?: string
  yo?: string
  bn?: string
  sr?: string
  sk?: string
  ta?: string
  cv?: string
  ce?: string
  hy?: string
  ca?: string
  ur?: string
  mn?: string
  os?: string
  jv?: string
  mk?: string
  vi?: string
  id?: string
  nn?: string
  wo?: string
  ja?: string
  vo?: string
  hu?: string
  ro?: string
  af?: string
  da?: string
  tr?: string
  so?: string
  pa?: string
  ml?: string
  eo?: string
  nl?: string
  kk?: string
  fa?: string
  io?: string
  hr?: string
  zh?: string
  gd?: string
  ht?: string
  br?: string
  uz?: string
  az?: string
  ky?: string
  sc?: string
  bs?: string
  th?: string
  ie?: string
  tl?: string
  ku?: string
  no?: string
  el?: string
  ka?: string
  sv?: string
  gn?: string
  la?: string
  bo?: string
  mr?: string
  pt?: string
  hi?: string
  sl?: string
}